Moved around some classes and packages
