--- src/main/java/library/commandrunner/io/reader/InputReader.java
@@ -1,4 +1,0 @@
-package library.commandrunner.io.reader;
-
-public interface InputReader {
-}
