+++ src/main/java/library/commandrunner/loan/LoanCommandRunner.java
@@ -7,7 +7,7 @@ import library.item.Item;
 import library.register.CustomerRegister;
 import library.register.Library;
 import library.util.Constant;
-import library.util.JSONConverter;
+import library.io.JSONConverter;
 
 import java.io.IOException;
 import java.util.List;
