+++ src/main/java/library/customer/Customer.java
@@ -8,8 +8,8 @@ import com.fasterxml.jackson.databind.annotation.JsonDeserialize;
 import com.fasterxml.jackson.databind.annotation.JsonSerialize;
 import library.util.Constant;
 import library.item.Borrowable;
-import library.util.LocalDateDeserializer;
-import library.util.LocalDateSerializer;
+import library.io.reader.LocalDateDeserializer;
+import library.io.writer.LocalDateSerializer;
 
 import java.time.LocalDate;
 import java.util.ArrayList;
