+++ src/main/java/library/io/reader/InputReader.java
@@ -1,0 +1,4 @@
+package library.io.reader;
+
+public interface InputReader {
+}
