+++ src/main/java/library/io/writer/LocalDateSerializer.java
@@ -1,4 +1,4 @@
-package library.util;
+package library.io.writer;
 
 import com.fasterxml.jackson.core.JsonGenerator;
 import com.fasterxml.jackson.databind.SerializerProvider;
