+++ src/main/java/library/register/CustomerRegister.java
@@ -1,7 +1,7 @@
 package library.register;
 
 import library.customer.Customer;
-import library.util.JSONConverter;
+import library.io.JSONConverter;
 
 import java.util.ArrayList;
 import java.util.List;
